# doc/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SoNNMT Document Router",
    "version": "2.0.0"
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/ready": {
      "get": {
        "summary": "Ready",
        "operationId": "ready_ready_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/v1/directory": {
      "get": {
        "summary": "Directory",
        "operationId": "directory_v1_directory_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/v1/jobs": {
      "post": {
        "summary": "Create Job",
        "operationId": "create_job_v1_jobs_post",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "idempotency-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 128
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_job_v1_jobs_post"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/jobs/{job_id}": {
      "get": {
        "summary": "Get Job",
        "operationId": "get_job_v1_jobs__job_id__get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/classify": {
      "post": {
        "summary": "Classify",
        "description": "Compatibility endpoint. Prefer jobs for PDF/OCR to avoid client timeouts.",
        "operationId": "classify_classify_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_classify_classify_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClassifyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/v1/jobs/{job_id}/feedback": {
      "post": {
        "summary": "Feedback",
        "operationId": "feedback_v1_jobs__job_id__feedback_post",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          },
          {
            "name": "idempotency-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 128
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Feedback"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Body_classify_classify_post": {
        "properties": {
          "document_id": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Document Id",
            "default": "legacy"
          },
          "received_on": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Received On"
          },
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          },
          "file": {
            "anyOf": [
              {
                "type": "string",
                "contentMediaType": "application/octet-stream"
              },
              {
                "type": "null"
              }
            ],
            "title": "File"
          }
        },
        "type": "object",
        "title": "Body_classify_classify_post"
      },
      "Body_create_job_v1_jobs_post": {
        "properties": {
          "document_id": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Document Id"
          },
          "received_on": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Received On"
          },
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          },
          "files": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "contentMediaType": "application/octet-stream"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Files"
          },
          "file": {
            "anyOf": [
              {
                "type": "string",
                "contentMediaType": "application/octet-stream"
              },
              {
                "type": "null"
              }
            ],
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "document_id"
        ],
        "title": "Body_create_job_v1_jobs_post"
      },
      "ClassifyResponse": {
        "properties": {
          "don_vi_xu_ly_chinh": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 40,
            "title": "Don Vi Xu Ly Chinh"
          },
          "phoi_hop_xu_ly": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 40,
            "title": "Phoi Hop Xu Ly"
          },
          "lanh_dao_theo_doi": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 40,
            "title": "Lanh Dao Theo Doi"
          },
          "han_thuc_hien": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Han Thuc Hien"
          },
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "document_id": {
            "type": "string",
            "title": "Document Id"
          },
          "input_sha256": {
            "type": "string",
            "title": "Input Sha256"
          },
          "rules_version": {
            "type": "string",
            "title": "Rules Version"
          },
          "rules_sha256": {
            "type": "string",
            "title": "Rules Sha256"
          },
          "directory_version": {
            "type": "string",
            "title": "Directory Version"
          },
          "received_on": {
            "type": "string",
            "format": "date",
            "title": "Received On"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "default": 0
          },
          "confidence_kind": {
            "type": "string",
            "const": "heuristic_uncalibrated",
            "title": "Confidence Kind",
            "default": "heuristic_uncalibrated"
          },
          "reason": {
            "type": "string",
            "title": "Reason",
            "default": ""
          },
          "matched_rules": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Matched Rules"
          },
          "needs_review": {
            "type": "boolean",
            "title": "Needs Review",
            "default": true
          },
          "requires_confirmation": {
            "type": "boolean",
            "title": "Requires Confirmation",
            "default": true
          },
          "degraded": {
            "type": "boolean",
            "title": "Degraded",
            "default": false
          },
          "tier": {
            "type": "string",
            "enum": [
              "T2",
              "T1",
              "T0"
            ],
            "title": "Tier",
            "default": "T0"
          },
          "do_khan": {
            "type": "string",
            "enum": [
              "thuong",
              "khan",
              "thuong_khan",
              "hoa_toc"
            ],
            "title": "Do Khan",
            "default": "thuong"
          },
          "review_reasons": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Review Reasons"
          },
          "extracted_metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Extracted Metadata"
          },
          "evidence": {
            "additionalProperties": true,
            "type": "object",
            "title": "Evidence"
          },
          "extraction": {
            "additionalProperties": true,
            "type": "object",
            "title": "Extraction"
          },
          "recipients": {
            "additionalProperties": {
              "items": {
                "$ref": "#/components/schemas/Recipient"
              },
              "type": "array"
            },
            "type": "object",
            "title": "Recipients"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "job_id",
          "document_id",
          "input_sha256",
          "rules_version",
          "rules_sha256",
          "directory_version",
          "received_on"
        ],
        "title": "ClassifyResponse"
      },
      "Feedback": {
        "properties": {
          "document_id": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Document Id"
          },
          "input_sha256": {
            "type": "string",
            "title": "Input Sha256"
          },
          "rules_sha256": {
            "type": "string",
            "title": "Rules Sha256"
          },
          "decision": {
            "type": "string",
            "enum": [
              "accepted",
              "edited",
              "rejected"
            ],
            "title": "Decision"
          },
          "final": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RoutingFields"
              },
              {
                "type": "null"
              }
            ]
          },
          "do_khan": {
            "type": "string",
            "enum": [
              "thuong",
              "khan",
              "thuong_khan",
              "hoa_toc"
            ],
            "title": "Do Khan",
            "default": "thuong"
          },
          "note": {
            "type": "string",
            "maxLength": 2000,
            "title": "Note",
            "default": ""
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "document_id",
          "input_sha256",
          "rules_sha256",
          "decision"
        ],
        "title": "Feedback"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "JobStatus": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "completed",
              "failed"
            ],
            "title": "Status"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ClassifyResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "status"
        ],
        "title": "JobStatus"
      },
      "Recipient": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "Recipient"
      },
      "RoutingFields": {
        "properties": {
          "don_vi_xu_ly_chinh": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 40,
            "title": "Don Vi Xu Ly Chinh"
          },
          "phoi_hop_xu_ly": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 40,
            "title": "Phoi Hop Xu Ly"
          },
          "lanh_dao_theo_doi": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 40,
            "title": "Lanh Dao Theo Doi"
          },
          "han_thuc_hien": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Han Thuc Hien"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "RoutingFields"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  }
}